app/requirements.txt
asyncio==3.4.3
bcrypt==4.3.0
fastapi==0.104.1
jose==1.0.0
passlib==1.7.4
pydantic==2.11.3
pydantic_core==2.33.1
python-dotenv==1.1.1
requests==2.32.4
shortuuid==1.0.13
uvicorn==0.24.0